requirements.txt
firebase-admin==6.5.0
paho-mqtt==1.6.1
